Added network stats

+++ lib/functions/memory_conversion.dart
@@ -8,6 +8,6 @@ String convertMemoryToMb(double value) {
   return (value/pow(1024, 2)).toStringAsFixed(2);
 }
 
-String convertMemoryToKb(int value) {
+String convertMemoryToKb(double value) {
   return (value/1024).toStringAsFixed(2);
 }
